Split cron tasks between server/dump-manager.
