+++ lsif/src/dump-manager/dump-manager.ts
@@ -7,6 +7,9 @@ import { ensureDirectory } from '../shared/paths'
 import { Logger } from 'winston'
 import { startExpressApp } from '../shared/api/init'
 import { createDatabaseRouter } from './routes/database'
+import { createPostgresConnection } from '../shared/database/postgres'
+import { waitForConfiguration } from '../shared/config/config'
+import { startTasks } from './tasks'
 
 /**
  * No-op dump-manager process.
@@ -17,11 +20,20 @@ async function main(logger: Logger): Promise<void> {
     // Collect process metrics
     promClient.collectDefaultMetrics({ prefix: 'lsif_' })
 
+    // Read configuration from frontend
+    const fetchConfiguration = await waitForConfiguration(logger)
+
     // Ensure storage roots exist
     await ensureDirectory(settings.STORAGE_ROOT)
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.DBS_DIR))
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.TEMP_DIR))
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.UPLOADS_DIR))
+
+    // Create database connection
+    const connection = await createPostgresConnection(fetchConfiguration(), logger)
+
+    // Start background tasks
+    startTasks(connection, logger)
 
     const routers = [createDatabaseRouter(logger)]
