+++ lsif/src/dump-manager/settings.ts
@@ -18,3 +18,27 @@ export const DOCUMENT_CACHE_CAPACITY = readEnvInt('DOCUMENT_CACHE_CAPACITY', 102
 
 /** The maximum number of result chunks that can be held in memory at once. */
 export const RESULT_CHUNK_CACHE_CAPACITY = readEnvInt('RESULT_CHUNK_CACHE_CAPACITY', 1024 * 1024 * 1024)
+
+/** How many uploads to query at once when determining if a db file is unreferenced. */
+export const DEAD_DUMP_CHUNK_SIZE = readEnvInt('DEAD_DUMP_CHUNK_SIZE', 100)
+
+/** The maximum age (in seconds) that the files for an unprocessed upload can remain on disk. */
+export const FAILED_UPLOAD_MAX_AGE = readEnvInt('FAILED_UPLOAD_MAX_AGE', 24 * 60 * 60)
+
+/** The interval (in seconds) to clean the dbs directory. */
+export const PURGE_OLD_DUMPS_INTERVAL = readEnvInt('PURGE_OLD_DUMPS_INTERVAL', 60 * 30)
+
+/** The interval (in seconds) to invoke the cleanFailedUploads task. */
+export const CLEAN_FAILED_UPLOADS_INTERVAL = readEnvInt('CLEAN_FAILED_UPLOADS_INTERVAL', 60 * 60 * 8)
+
+/** The maximum space (in bytes) that the dbs directory can use. */
+export const DBS_DIR_MAXIMUM_SIZE_BYTES = readEnvInt('DBS_DIR_MAXIMUM_SIZE_BYTES', 1024 * 1024 * 1024 * 10)
+
+/** How many times to retry requests to lsif-server in the background. */
+export const MAX_REQUEST_RETRIES = readEnvInt('MAX_REQUSET_RETRIES', 60)
+
+/** How long to wait (in seconds) between lsif-server request attempts. */
+export const REQUEST_RETRY_INTERVAL = readEnvInt('REQUSET_RETRY_INTERVAL', 5)
+
+/** HTTP address for internal LSIF HTTP API. */
+export const LSIF_SERVER_URL = process.env.LSIF_SERVER_URL || 'http://lsif-server'
